+++ react/features/conference/middleware.native.ts
@@ -1,0 +1,1 @@
+import './middleware.any';
